splitup raster_legend
